produce tables from xlsx data dictionary

produce tables from xlsx data dictionary

process_datadict(path, quiet = TRUE, simplify_names = TRUE)

Arguments

path

character(1) path to xlsx file with data dictionary in several sheets

quiet

logical(1) if TRUE (default) suppress messages about column names from read_xlsx

simplify_names

logical(1) if TRUE (default) remove inessential elements of sheet names

Value

list of data.frames

Note

Original xlsx file retrieved from https://data.cdc.gov/api/views/unsk-b7fc/files/bfb13090-4148-4f48-9d5a-e1d1c2fb93b3?download=true&filename=DataDictionary_v36_12082022.xlsx in early 2023.

Examples

pa = system.file("cdc/VACCDataDictionary_v36_12082022.xlsx", package="teachCovidData")
pd = process_datadict(pa)
names(pd)
#>  [1] "0. Notes"                        "1. Vaccinations US Jurisdiction"
#>  [3] "2. Vaccinations US Trends"       "3. Vaccinations US Demograp"    
#>  [5] "4. Age Sex Trends"               "5. Vaccinations US County"      
#>  [7] "6. CaseTrends AgeGp"             "7. Booster Dose Eligibility"    
#>  [9] "8. Primary and Booster Chart"    "9. Jurisdiction Abbreviations"  
head(pd[[1]])
#>                                   CDC COVID-19 Vaccine Administration and Distribution data
#> 1                                                      Recent as of 11/17/2022 @ 8:00 AM ET
#> 2                                                   Historical data available for download:
#> 3                                  COVID-19 Vaccinations in the United States, Jurisdiction
#> 4             COVID-19 Vaccination Trends in the United States, National and Jurisdictional
#> 5 COVID-19 Vaccination Age and Sex Trends in the United States, National and Jurisdictional
#> 6                                  COVID-19 Vaccinations in the United States, County level
#>   ...2                                    ...3
#> 1 <NA>                                    <NA>
#> 2 <NA> Associated CDC COVID Data Tracker Site:
#> 3    →       Vaccinations in the United States
#> 4    →                      Vaccination Trends
#> 5    →          Vaccination Demographic Trends
#> 6    →                  Vaccinations by County